+++ packages/nexrender-core/src/tasks/script/wrap-enhanced-script.js
@@ -1,48 +1,4 @@
 const fs     = require('fs')
-const path   = require('path')
-const script = require('../assets/nexrender.jsx')
-const { checkForWSL } = require('../helpers/path')
-
-/* helpers */
-const escape = str => {
-    str = JSON.stringify(str)
-    str = str.substring(1, str.length-1)
-    str = `'${str.replace(/'/g, '\\\'')}'`
-    return str
-}
-
-const selectLayers = (job, settings, { composition, layerName, layerIndex }, callbackString) => {
-    const method = layerName ? 'selectLayersByName' : 'selectLayersByIndex';
-    const compo  = composition === undefined ? 'null' : escape(composition);
-    const value  = layerName ? escape(layerName) : layerIndex;
-    return (`nexrender.${method}(${compo}, ${value}, ${callbackString}, null, ${job.template.continueOnMissing});`);
-}
-
-const renderIf = (value, string) => {
-    const encoded = typeof value == 'string' ? escape(value) : JSON.stringify(value);
-    return value === undefined ? '' : string.replace('$value', () => encoded);
-}
-
-const partsOfKeypath = (keypath) => {
-    var parts = keypath.split('->');
-    return (parts.length === 1) ? keypath.split('.') : parts
-}
-
-/* scripting wrappers */
-const wrapFootage = (job, settings, { dest, ...asset }) => (`(function() {
-    ${selectLayers(job, settings, asset, `function(layer) {
-        nexrender.replaceFootage(layer, '${checkForWSL(dest.replace(/\\/g, "\\\\"), settings)}')
-    }`)}
-})();\n`)
-
-const wrapData = (job, settings, { property, value, expression, ...asset }) => (`(function() {
-    ${selectLayers(job, settings, asset, `function(layer) {
-        var parts = ${JSON.stringify(partsOfKeypath(property))};
-        ${renderIf(value, `var value = { "value": $value }`)}
-        ${renderIf(expression, `var value = { "expression": $value }`)}
-        nexrender.changeValueForKeypath(layer, parts, value);
-    }`)}
-})();\n`)
 
 /**
  *   Wrap Enhanced Script
@@ -428,47 +384,4 @@ const wrapEnhancedScript = (job, settings, { dest, src, parameters = [], keyword
     return enhancedScript.build();
 }
 
-module.exports = (job, settings) => {
-    settings.logger.log(`[${job.uid}] running script assemble...`);
-
-    const data = [];
-    const base = job.workpath;
-
-    job.assets.map(asset => {
-        settings.trackCombined('Asset Script Wraps', {
-            job_id: job.uid, // anonymized internally
-            script_type: asset.type,
-            script_compostion_set: asset.composition !== undefined,
-            script_layer_strat: asset.layerName ? 'name' : 'index',
-            script_value_strat:
-                asset.value !== undefined ? 'value' : // eslint-disable-line no-nested-ternary, multiline-ternary
-                asset.expression !== undefined ? 'expression' : // eslint-disable-line multiline-ternary
-                undefined,
-        })
-
-        switch (asset.type) {
-            case 'video':
-            case 'audio':
-            case 'image':
-                data.push(wrapFootage(job, settings, asset));
-                break;
-
-            case 'data':
-                data.push(wrapData(job, settings, asset));
-                break;
-
-            case 'script':
-                data.push(wrapEnhancedScript(job, settings, asset));
-                break;
-        }
-    });
-
-    /* write out assembled custom script file in the workpath */
-    job.scriptfile = path.join(base, `nexrender-${job.uid}-script.jsx`);
-    fs.writeFileSync(job.scriptfile, script
-        .replace('/*COMPOSITION*/', job.template.composition)
-        .replace('/*USERSCRIPT*/', () => data.join('\n'))
-    );
-
-    return Promise.resolve(job)
-}
+module.exports = wrapEnhancedScript
